Fix: remove duplicated test and fix test
